Let user enable/disable C++ and Fortran support
Expose two new configuration variables (which default to yes):
HDF5_ENABLE_CXX and HDF5_ENABLE_FORTRAN setting them to no allows
users to avoid building/linking against the respective interfaces to
hdf5. This is useful when eg. one does not need them and would like to
link against a cluster supplied version of hdf5 (or if one cannot use
the Fortran modules due to differing compiler versions). This is
related to the discussion in
http://lists.einsteintoolkit.org/pipermail/users/2011-March/000928.html
although it does not provide any early warning.

Patch by Roland Haas


git-svn-id: http://svn.cactuscode.org/projects/ExternalLibraries/HDF5/trunk@38 fb53df36-e548-4a1e-8150-ab98cbd5e786